requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
bitsandbytes==0.44.1
braceexpand==0.1.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.4
imageio==2.36.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.6.1
joblib==1.4.2
jupyter-client==7.3.4
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nest_asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.2
numba==0.60.0
numpy==1.26.3
ollama==0.3.3
omegaconf==2.3.0
openai==1.61.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.2
platformdirs==4.3.6
portalocker==3.1.1
progressbar2==4.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0
python-utils==3.9.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.9.11
requests==2.28.1
safetensors==0.4.5
scikit-dimension==0.3.4
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.2.0
sentencepiece==0.2.0
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==75.1.0
shortuuid==1.0.13
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.5.0
tifffile==2025.3.30
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.15.2
tornado==6.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.37.2
transformers-stream-generator==0.0.5
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.13
visual-genome==1.1.1
wandb==0.19.9
wcwidth==0.2.13
webdataset==0.2.111
wheel==0.44.0
yarl==1.15.3
